分类相关接口业务开发

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,17 +2,22 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 
@@ -29,6 +34,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    //菜品类
+    @Autowired
+    private DishMapper dishMapper;
+
+    //套餐类
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     /**
      * 菜品分类分页查询
@@ -86,12 +99,25 @@ public class CategoryServiceImpl implements CategoryService {
      *
      * @param id
      */
+    // TODO 当第二页只有一个类别时，将其删除，页面查询没有任何数据（待处理）
+    @Transactional
     public void removeCategory(Long id) {
+        //需要查询当前类别下是否还有菜品，如果没有才能删除
+        Long countDish = dishMapper.countByCategoryId(id);
+        //查询当前分类是否关联了菜品，如果关联了就抛出业务异常
+        if (countDish > 0) throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+
+        //查询当前分类是否关联了套餐，如果关联了就抛出业务异常
+        Long countSetmeal = setmealMapper.countByCategoryId(id);
+        if (countSetmeal > 0) throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+
+        //判断结束，进行删除
         categoryMapper.deleteCategory(id);
     }
 
     /**
      * 根据类型查询
+     *
      * @param type
      * @return
      */
@@ -101,12 +127,13 @@ public class CategoryServiceImpl implements CategoryService {
 
     /**
      * 修改菜品分类
+     *
      * @param categoryDTO
      */
     public void edit(CategoryDTO categoryDTO) {
         //对象拷贝到Category对象中
         Category category = new Category();
-        BeanUtils.copyProperties(categoryDTO,category);
+        BeanUtils.copyProperties(categoryDTO, category);
         //修改最后操作时间和最后操作用户
         Long id = BaseContext.getCurrentId();
         category.setUpdateTime(LocalDateTime.now());
